--- a/src/Genie.jl
+++ b/src/Genie.jl
@@ -3,6 +3,11 @@ Loads dependencies and bootstraps a Genie app. Exposes core Genie functionality.
 """
 module Genie
 
+import EzXML
+import FilePathsBase
+import HTTP
+import Markdown
+import Pkg
 import Revise
 
 push!(LOAD_PATH, @__DIR__)
@@ -18,9 +23,9 @@ include("constants.jl")
 
 import Sockets
 import Logging
-import PrecompileSignatures
 
 using Reexport
+using PrecompileSignatures: @precompile_signatures
 
 include(joinpath(@__DIR__, "genie_types.jl"))
 
@@ -286,14 +291,6 @@ end
 
 const bootstrap = genie
 
-import EzXML
-import FilePathsBase
-import Markdown
-import HTTP
-import Pkg
-
-if ccall(:jl_generating_output, Cint, ()) == 1
-    include(PrecompileSignatures.precompile_directives(Genie))
-end
+@precompile_signatures(Genie)
 
 end