I wrote a package called [`PrecompileSignatures.jl`](https://github.com/rikhuijzer/PrecompileSignatures.jl) to speed up package loading time. When adding it to `Genie.jl`, it makes `precompile` take 16 seconds longer, makes `using Genie` take 1 second longer and reduces the time for the first running webserver by 1 second. I leave it up to you to decide whether this tradeoff is interesting/worth it. Note that I only benchmarked the code below, so there might be more effect for some other function calls.

## Benchmark

```julia
using Pkg
Pkg.precompile()

@time @eval using Genie
import Genie.Router: route
import Genie.Renderer.Json: json

using Genie.HTTP

# Warmup
HTTP.get("https://example.com")

@async begin
    Genie.config.run_as_server = true

    route("/") do
        (:message => "Hi there!") |> json
    end

    Genie.startup()
end

# Give the server some time to start.
sleep(1)

@time @eval HTTP.get("http://127.0.0.1:8000")
```

### Before this PR (`master`) - Julia 1.7.3

```
  1 dependency successfully precompiled in 8 seconds (55 already precompiled)
  3.150201 seconds (5.92 M allocations: 342.803 MiB, 6.34% gc time, 67.92% compilation time)
  7.201561 seconds (16.62 M allocations: 873.522 MiB, 4.33% gc time, 99.97% compilation time)
```

### This PR - Julia 1.7.3

```
1 dependency successfully precompiled in 25 seconds (55 already precompiled)
  4.375188 seconds (8.38 M allocations: 500.447 MiB, 6.07% gc time, 49.09% compilation time)
  6.329759 seconds (14.37 M allocations: 755.809 MiB, 3.84% gc time, 99.96% compilation time)
```