@rikhuijzer 2022-04-29
@essenciary I'll contact you on Slack/Zulip to schedule a meeting. I think that's the quickest way to answer your questions