@essenciary 2022-05-11
It doesn't seem to have the same effect as precompiling a custom image... how can we confirm that things get done? 

<img width="1214" alt="image" src="https://user-images.githubusercontent.com/11292378/167866115-b8a5f771-abd1-4420-989a-dff9fcb4d1e8.png">