@essenciary 2022-05-11
Very interesting, with a precompiled image, RAM usage hover a bit over 500 MB. While with non-precompiled it's over 700 MB, flat.