@rikhuijzer 2022-05-11
> It doesn't seem to have the same effect as precompiling a custom image... how can we confirm that things get done?

That's correct. `PackageCompiler.jl` will have a greater effect. `PrecompileSignatures.jl` has less effects on reducing TTFX, but is easier to integrate in a package.