@essenciary 2022-05-12
> > It doesn't seem to have the same effect as precompiling a custom image... how can we confirm that things get done?
> 
> That's correct. `PackageCompiler.jl` will have a greater effect. `PrecompileSignatures.jl` has less effects on reducing TTFX, but is easier to integrate in a package.

Thanks Rik! Is there any way to check/measure if/how much was precompiled? It seems to go very fast over that block of code.