@rikhuijzer 2022-05-12
> Is there any way to check/measure if/how much was precompiled?

You can verify whether method `f` was precompiled by checking `methods(f)[1].specializations`. If it is precompiled, then there will be specializations. More generally, I'm just running benchmarks `@time @eval begin` to see the effects and don't know easier ways to check the effects.