@rikhuijzer 2022-05-12
I think I misunderstood your question. I've added the following to the [README.md](https://github.com/rikhuijzer/PrecompileSignatures.jl) which might answer your question:

> ## How can I see what signatures this package creates?
>
> Use
> ```julia
> PrecompileSignatures.precompilables(Foo)
> ```
> where `Foo` is the name of your package.
>
> ## How can I see whether the created signatures are valid?
>  
> By default, this package doesn't verify the `precompile` directives to avoid breaking precompilation in worst-case scenarios.
> To verify the directives for debugging purposes, use
> ```julia
> types = PrecompileSignatures.precompilables(Foo)
> @assert all(precompile.(types))
> ```
> where `Foo` is the name of your package.